feat: make one generic ModificationDialog
